+++ wp-content/plugins/wordpress-beta-tester/CHANGES.md
@@ -1,5 +1,28 @@
 [unreleased]
 
+#### 2.2.8 / 2020-03-17 🍀
+* add Dev Notes and Field Guide links to dashboard
+* add text/link for bug reporting to trac
+* add help tabs to screen
+* arbitrarily changed settings page id from `wp_beta_tester` to `wp-beta-tester` 😏
+
+#### 2.2.7 / 2020-03-02
+* update trac link in callout for _closed_ or _reopened_ tickets on the milestone
+* only show Beta Tester Settings page link in callout with appropriate privileges, using `manage_network_options` and `manage_options`
+* menu to Settings page also checks privileges as above
+
+#### 2.2.6 / 2020-02-25
+* removed extra `</li>` in dashboard callout, 4th time's the charm 😭
+
+#### 2.2.5 / 2020-02-25
+* less greedy regex for matching release posts in RSS for dashboard callout
+
+#### 2.2.4 / 2020-02-25 🤦‍♂️
+* added dashboard widget for network dashboard
+
+#### 2.2.3 / 2020-02-25
+* add dashboard widget callout for testing
+
 #### 2.2.2 / 2020-02-22
 * fix for strange Core API response where preferred version response contained the word 'version'. We now grab the last word of that response